Code clean up: - #1581
